+++ src/packet/key/symmetric.rs
@@ -6,13 +6,19 @@
 
 use std::{fmt::Debug, io::BufRead};
 
+use aead::rand_core::CryptoRng;
+use bytes::Bytes;
+use elliptic_curve::subtle::ConstantTimeEq;
 use log::debug;
-use rand::thread_rng;
+use rand::{thread_rng, Rng};
 
 use crate::{
     composed::PlainSessionKey,
-    crypto::{aead::AeadAlgorithm, hash::HashAlgorithm, public_key::PublicKeyAlgorithm},
-    errors::{bail, ensure_eq, unsupported_err},
+    crypto::{
+        aead::AeadAlgorithm, aead_key::InfoParameter, hash::HashAlgorithm,
+        public_key::PublicKeyAlgorithm,
+    },
+    errors::{bail, ensure, ensure_eq, unsupported_err, Error},
     packet::{PacketHeader, PacketTrait, PubKeyInner, SignatureVersion},
     ser::Serialize,
     types::{
@@ -92,6 +98,118 @@ impl PersistentSymmetricKey {
             details,
             secret_params,
         })
+    }
+
+    pub(crate) fn symmetric_encrypt<R: CryptoRng + Rng>(
+        &self,
+        mut rng: R,
+        pw: &Password,
+        plain: &[u8],
+        typ: EskType,
+        aead: AeadAlgorithm,
+        version: KeyVersion,
+    ) -> Result<PkeskBytes, Error> {
+        ensure!(
+            matches!(typ, EskType::V6),
+            "only v6 ESK supported right now"
+        );
+
+        self.unlock(pw, |pub_params, sec_params| {
+            let PublicParams::AEAD(public_params) = pub_params else {
+                bail!("Unsupported public parameters for persistent symmetric key: {pub_params:?}");
+            };
+
+            let PlainSecretParams::AEAD(secret) = &sec_params else {
+                bail!("Unsupported secret parameters for persistent symmetric key: {sec_params:?}");
+            };
+
+            // 32 octets of salt. The salt is used to derive the key-encryption key and MUST be
+            // securely generated (see section 13.10 of [RFC9580]).
+            let mut salt: [u8; 32] = [0; 32];
+            rng.fill(&mut salt);
+
+            // A symmetric key encryption of the plaintext value described in section 5.1 of [RFC9580],
+            // performed with the key-encryption key and IV computed as described in Section 7.4,
+            // using the symmetric-key cipher of the key and the indicated AEAD mode, with as
+            // additional data the empty string; including the authentication tag.
+
+            let version = version.into();
+            let info = InfoParameter {
+                packet_type: Tag::PublicKeyEncryptedSessionKey,
+                version,
+                aead,
+                sym_alg: public_params.sym_alg,
+            };
+
+            let (key, iv) =
+                crate::crypto::aead_key::SecretKey::derive_key_iv(&secret.key, &salt, info);
+
+            let mut buf = plain.into();
+
+            aead.encrypt_in_place(&public_params.sym_alg, &key, &iv, &[], &mut buf)?;
+
+            let encrypted: Bytes = buf.into();
+
+            Ok(PkeskBytes::Aead {
+                aead,
+                salt,
+                encrypted,
+            })
+        })?
+    }
+
+    pub(crate) fn symmetric_verify(
+        &self,
+        pw: &Password,
+        hash: HashAlgorithm,
+        data: &[u8],
+        sig: &SignatureBytes,
+    ) -> Result<(), Error> {
+        let Some(digest_len) = hash.digest_size() else {
+            bail!(
+                "UnlockablePersistentSymmetricKey::verify: invalid hash algorithm: {:?}",
+                hash
+            );
+        };
+        ensure_eq!(
+            data.len(),
+            digest_len,
+            "signature data length {} doesn't match digest len {}",
+            data.len(),
+            digest_len,
+        );
+
+        let SignatureBytes::PersistentSymmetric { aead, salt, tag } = sig else {
+            bail!("Unsupported SignatureBytes for persistent symmetric key: {sig:?}");
+        };
+
+        ensure_eq!(
+            tag.len(),
+            aead.tag_size().unwrap_or(0),
+            "unexpected tag length"
+        );
+
+        self.unlock(pw, |pub_params, sec_params| {
+            let PublicParams::AEAD(public) = &pub_params else {
+                bail!("Unsupported public parameters for persistent symmetric key: {pub_params:?}");
+            };
+            let PlainSecretParams::AEAD(secret) = &sec_params else {
+                bail!("Unsupported secret parameters for persistent symmetric key: {sec_params:?}");
+            };
+
+            let version = SignatureVersion::V6; // FIXME: should not be fixed
+
+            // "buf" is the newly calculated authentication tag
+            let buf = secret.compute_persistent_mac(version, public.sym_alg, *aead, salt, data)?;
+
+            // check if the stored and calculated authentication tags match
+            if buf.ct_ne(&**tag).into() {
+                // no: the signature is invalid!
+                bail!("PersistentSymmetricKey signature mismatch");
+            }
+
+            Ok(())
+        })?
     }
 
     pub fn secret_params(&self) -> &SecretParams {
